requirements.txt
pandas>=2.0
numpy>=1.24
nltk>=3.8
scikit-learn>=1.3
vaderSentiment>=3.3.2
matplotlib>=3.7
seaborn>=0.13
transformers>=4.36
torch>=2.1
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
python-dotenv>=1.0
jupyter>=1.0
ipykernel>=6.29